+++ modules/devices.py
@@ -98,6 +98,17 @@ def test_fp16():
         shared.opts.no_half_vae = True
         return False
 
+def test_bf16():
+    if shared.cmd_opts.experimental:
+        return True
+    try:
+        import torch.nn.functional as F
+        image = torch.randn(1, 4, 32, 32).to(device="cuda", dtype=torch.bfloat16)
+        _out = F.interpolate(image, size=(64, 64), mode="nearest")
+    except:
+        shared.log.warning('Torch BF16 test failed: Fallback to FP16 operations')
+        return False
+
 
 def set_cuda_params():
     shared.log.debug('Verifying Torch settings')
@@ -117,25 +128,32 @@ def set_cuda_params():
             except:
                 pass
     global dtype, dtype_vae, dtype_unet, unet_needs_upcast # pylint: disable=global-statement
-    ok = test_fp16()
     if shared.cmd_opts.use_directml and not shared.cmd_opts.experimental: # TODO DirectML does not have full autocast capabilities
         shared.opts.no_half = True
         shared.opts.no_half_vae = True
-    if ok and shared.opts.cuda_dtype == 'FP32':
-        shared.log.info('CUDA FP16 test passed but desired mode is set to FP32')
-    if shared.opts.cuda_dtype == 'FP16' and ok:
-        dtype = torch.float16
-        dtype_vae = torch.float16
-        dtype_unet = torch.float16
-    if shared.opts.cuda_dtype == 'BF16' and ok:
-        dtype = torch.bfloat16
-        dtype_vae = torch.bfloat16
-        dtype_unet = torch.bfloat16
-    if shared.opts.cuda_dtype == 'FP32' or shared.opts.no_half or not ok:
+    if shared.opts.cuda_dtype == 'FP32':
+        dtype = torch.float32
+        dtype_vae = torch.float32
+        dtype_unet = torch.float32
+    if shared.opts.cuda_dtype == 'BF16' or dtype == torch.bfloat16:
+        bf16_ok = test_bf16()
+        dtype = torch.bfloat16 if bf16_ok else torch.float16
+        dtype_vae = torch.bfloat16 if bf16_ok else torch.float16
+        dtype_unet = torch.bfloat16 if bf16_ok else torch.float16
+    if shared.opts.cuda_dtype == 'FP16' or dtype == torch.bfloat16:
+        fp16_ok = test_fp16()
+        dtype = torch.float16 if fp16_ok else torch.float32
+        dtype_vae = torch.float16 if fp16_ok else torch.float32
+        dtype_unet = torch.float16 if fp16_ok else torch.float32
+    else:
+        pass
+    if shared.opts.no_half:
+        shared.log.info('Torch override dtype: no-half set')
         dtype = torch.float32
         dtype_vae = torch.float32
         dtype_unet = torch.float32
     if shared.opts.no_half_vae: # set dtype again as no-half-vae options take priority
+        shared.log.info('Torch override VAE dtype: no-half-vae set')
         dtype_vae = torch.float32
     unet_needs_upcast = shared.opts.upcast_sampling
     shared.log.debug(f'Desired Torch parameters: dtype={shared.opts.cuda_dtype} no-half={shared.opts.no_half} no-half-vae={shared.opts.no_half_vae} upscast={shared.opts.upcast_sampling}')
@@ -150,7 +168,7 @@ if args.use_ipex:
     CondFunc('torch.nn.modules.GroupNorm.forward',
         lambda orig_func, *args, **kwargs: orig_func(args[0], args[1].to(args[0].weight.data.dtype)),
         lambda *args, **kwargs: args[2].dtype != args[1].weight.data.dtype)
-    
+
     #Use XPU instead of CPU. %20 Perf improvement on weak CPUs.
     if args.device_id is not None:
         cpu = torch.device(f"xpu:{args.device_id}")