+++ modules/ui_extra_networks.py
@@ -4,7 +4,7 @@ import os.path
 import urllib.parse
 from pathlib import Path
 import gradio as gr
-from modules import shared
+from modules import shared, scripts
 from modules.generation_parameters_copypaste import image_from_url_text
 from modules.ui_components import ToolButton
 
@@ -54,9 +54,26 @@ class ExtraNetworksPage:
         self.card_short = shared.html("extra-networks-card-short.html")
         self.allow_negative_prompt = False
         self.metadata = {}
+        self.items = []
 
     def refresh(self):
         pass
+
+    def create_xyz_grid(self):
+        xyz_grid = [x for x in scripts.scripts_data if x.script_class.__module__ == "xyz_grid.py"][0].module
+
+        def add_prompt(p, opt, x):
+            for item in [x for x in self.items if x["name"] == opt]:
+                try:
+                    p.prompt = f'{p.prompt} {eval(item["prompt"])}' # pylint: disable=eval-used
+                except Exception as e:
+                    shared.log.error(f'Cannot evaluate extra network prompt: {item["prompt"]} {e}')
+
+        if not any(self.title in x.label for x in xyz_grid.axis_options):
+            if self.title == 'Checkpoints':
+                return
+            opt = xyz_grid.AxisOption(f"[Network] {self.title}", str, add_prompt, choices=lambda: [x["name"] for x in self.items])
+            xyz_grid.axis_options.append(opt)
 
     def link_preview(self, filename):
         quoted_filename = urllib.parse.quote(filename.replace('\\', '/'))
@@ -93,12 +110,14 @@ class ExtraNetworksPage:
         if subdirs:
             subdirs = {"": 1, **subdirs}
         subdirs_html = "".join([f"""
-<button class='lg secondary gradio-button custom-button{" search-all" if subdir=="" else ""}' onclick='extraNetworksSearchButton("{tabname}_extra_tabs", event)'>
-{html.escape(subdir if subdir!="" else "all")}
-</button>
-""" for subdir in subdirs])
+            <button class='lg secondary gradio-button custom-button{" search-all" if subdir=="" else ""}' onclick='extraNetworksSearchButton("{tabname}_extra_tabs", event)'>
+                {html.escape(subdir if subdir!="" else "all")}
+            </button>
+        """ for subdir in subdirs])
         try:
-            for item in self.list_items():
+            self.items = list(self.list_items())
+            self.create_xyz_grid()
+            for item in self.items:
                 metadata = item.get("metadata")
                 if metadata:
                     self.metadata[item["name"]] = metadata